Added upload of highscores to firebase
also fixed some datatypes since time is supposed to be a long since we are tracking it in milliseconds

+++ android/src/main/kotlin/com/blockshift/model/db/HighScore.kt
@@ -16,10 +16,10 @@ data class HighScore(
     val username: String,
     val levelID: Int,
     @ColumnInfo(name = "distance") val distance: Int,
-    @ColumnInfo(name = "time") val time: Int,
+    @ColumnInfo(name = "time") val time: Long,
     @ColumnInfo(name = "moves") val moves: Int
 )
 
 fun scoreIsZero(score:HighScore): Boolean {
-    return score.distance == 0 && score.time == 0 && score.moves == 0
+    return score.distance == 0 && score.time == 0L && score.moves == 0
 }
